+++ app/models/wrapper/TaxCategoryWrapper.java
@@ -1,4 +1,4 @@
-package models.export;
+package models.wrapper;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.neovisionaries.i18n.CountryCode;
